refactor[lotto]: 로또를 실행하는 기능 리팩토링(Application클래스 -> LottoGame클래스)
